Refactor keyspace response
